requirements.txt
absl-py==2.3.1
attrs==25.3.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
contourpy==1.3.0
cycler==0.12.1
filelock==3.19.1
flatbuffers==25.2.10
fonttools==4.59.1
fsspec==2025.7.0
hf-xet==1.1.7
huggingface-hub==0.34.4
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.6
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.4
mediapipe==0.10.9
ml_dtypes==0.5.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
opencv-contrib-python==4.11.0.86
opencv-python==4.12.0.88
opt_einsum==3.4.0
packaging==25.0
pandas==2.3.1
pillow==11.3.0
protobuf==3.20.3
psutil==7.0.0
py-cpuinfo==9.0.0
pycparser==2.22
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
safetensors==0.6.2
scipy==1.13.1
sentencepiece==0.2.1
six==1.17.0
sounddevice==0.5.2
sympy==1.14.0
timm==1.0.19
tokenizers==0.21.4
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.55.2
triton==3.4.0
typing_extensions==4.14.1
tzdata==2025.2
ultralytics==8.3.179
ultralytics-thop==2.0.15
urllib3==2.5.0
zipp==3.23.0
